requirements.txt
flask
gunicorn
qrcode
pillow
pytz
pymongo
python-dotenv